fix merge of branches for chowchow/chiwawa
Master branch has been restored to a linear history, but this patch apply
was failing:

commit 6f4d265c0a79fcd5411037a2dfd92c20b9c14a9d
Merge: 2c7eaed e7579ab
Author: m13 <mimmo@linino.org>
Date:   Mon Feb 23 01:04:28 2015 +0100

    Merge branch 'master' into chowchow2

    Conflicts:
        arch/mips/ath79/mach-linino.c
        arch/mips/include/asm/mach-ath79/mach-linino.h

I think it refers to a local master branch, not updated to github server,
so I can't easily understand what it means. I manually restored the
files as they are in chowchow2 branch.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>