AR933X uart driver: fixed values for scale and step, for most common baudrates
This patch gets rid of the #define UART_BAUDRATE_COMP_MEGA32U4, inside
ar933x_uart.c driver, and implements a more flexible way to feed the driver
with a set of predefined values for scale and step for the desired baudrates.
This permits to use the same driver for different chips (e.g. AR934X).
If baudrate is not in the list, the computation algorithm is performed.
A new Kconfig option, called TWEAK_UART_BAUDRATES_FOR_ATMEGA32U4, is
implemented for Atmega32u4 baudrates compatibility.

Signed-off-by: Aurelio Colosimo <aurelio@aureliocolosimo.it>